+++ login-workflow/src/components/ScrollViewWithBackground.tsx
@@ -6,14 +6,14 @@
 import React from 'react';
 
 // Components
-import { View, StyleSheet } from 'react-native';
+import { View, StyleSheet, ScrollViewProperties } from 'react-native';
 import { ScrollView } from 'react-native-gesture-handler';
+import { Theme, useTheme } from 'react-native-paper';
 
 /**
  * @ignore
  */
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-const makeStyles = (spaceSize: number, topColor: string, bottomColor: string) =>
+const makeStyles = (spaceSize: number, topColor: string, bottomColor: string): Record<string, any> =>
     StyleSheet.create({
         scrollViewTopCover: {
             backgroundColor: topColor,
@@ -36,10 +36,12 @@ const makeStyles = (spaceSize: number, topColor: string, bottomColor: string) =>
 /**
  * @param topColor  (Optional) The color of the top area's background.
  * @param bottomColor  (Optional) The color of the bottom area's background.
+ * @param theme (Optional) react-native-paper theme partial to style the component.
  */
-export type ScrollViewWithBackgroundProps = {
+export type ScrollViewWithBackgroundProps = ScrollViewProperties & {
     topColor?: string;
     bottomColor?: string;
+    theme?: Theme;
 };
 
 /**
@@ -48,11 +50,13 @@ export type ScrollViewWithBackgroundProps = {
  *
  * @category Component
  */
-export const ScrollViewWithBackground = ({ children, ...props }: any & ScrollViewWithBackgroundProps): JSX.Element => {
-    const styles = makeStyles(500, props.topColor ?? 'white', props.bottomColor ?? 'white');
+export const ScrollViewWithBackground: React.FC<ScrollViewWithBackgroundProps> = (props) => {
+    const { topColor, bottomColor, children, theme: customTheme, ...other } = props;
+    const theme = useTheme(customTheme);
+    const styles = makeStyles(500, topColor ?? theme.colors.surface, bottomColor ?? theme.colors.surface);
 
     return (
-        <ScrollView {...props}>
+        <ScrollView {...other}>
             <View style={styles.scrollViewTopCover} />
             <View style={styles.scrollViewBottomCover} />
             {children}