+++ login-workflow/src/components/Spinner.tsx
@@ -7,17 +7,13 @@ import React from 'react';
 
 // Components
 import { Platform, View, StyleSheet } from 'react-native';
-import { ActivityIndicator, Portal } from 'react-native-paper';
+import { ActivityIndicator, Portal, Theme, useTheme } from 'react-native-paper';
 import { getStatusBarHeight } from 'react-native-status-bar-height';
-
-// Styles
-import * as Colors from '@pxblue/colors';
 
 /**
  * @ignore
  */
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-const makeStyles = (hasHeader: boolean) =>
+const makeStyles = (hasHeader: boolean): Record<string, any> =>
     StyleSheet.create({
         overlay: {
             backgroundColor: 'rgba(255, 255, 255, 0.6)',
@@ -35,9 +31,11 @@ const makeStyles = (hasHeader: boolean) =>
 
 /**
  * @param hasHeader  If true, will allocate extra top space to not cover the header bar with the spinner.
+ * @param theme (Optional) react-native-paper theme partial to style the component.
  */
 type SpinnerProps = {
     hasHeader?: boolean;
+    theme?: Theme;
 };
 
 /**
@@ -45,14 +43,16 @@ type SpinnerProps = {
  *
  * @category Component
  */
-export function Spinner(props: SpinnerProps): JSX.Element {
-    const styles = makeStyles(props.hasHeader ?? true); // default has header is true
+export const Spinner: React.FC<SpinnerProps> = (props) => {
+    const { hasHeader = true } = props;
+    const theme = useTheme(props.theme);
+    const styles = makeStyles(hasHeader);
 
     return (
         <Portal>
             <View style={styles.overlay}>
-                <ActivityIndicator animating={true} color={Colors.blue['500']} style={styles.activityIndicator} />
+                <ActivityIndicator animating={true} color={theme.colors.primary} style={styles.activityIndicator} />
             </View>
         </Portal>
     );
-}
+};