+++ login-workflow/src/components/ToggleButton.tsx
@@ -8,7 +8,7 @@ import React from 'react';
 // Components
 import { StyleSheet, StyleProp, ViewStyle } from 'react-native';
 import { Label } from '@pxblue/react-native-components';
-import { Button } from 'react-native-paper';
+import { Button, Theme, useTheme } from 'react-native-paper';
 
 // Styles
 import * as Colors from '@pxblue/colors';
@@ -16,39 +16,40 @@ import * as Colors from '@pxblue/colors';
 /**
  * @ignore
  */
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-const makeStyles = (isOutlineOnly: boolean, isDisabled: boolean) =>
+const makeStyles = (props: ToggleButtonProps, theme: Theme): Record<string, any> =>
     StyleSheet.create({
         loginButton: {
             width: '100%',
             alignSelf: 'center',
-            borderRadius: 2,
-            borderColor: isOutlineOnly ? Colors.black['100'] : 'transparent',
+            borderRadius: theme.roundness,
+            borderColor: props.outlined ? Colors.black['100'] : 'transparent',
         },
         label: {
-            color: isOutlineOnly
-                ? isDisabled
+            color: props.outlined
+                ? props.disabled
                     ? Colors.black['100']
-                    : Colors.blue['500']
-                : isDisabled
+                    : theme.colors.primary
+                : props.disabled
                 ? Colors.gray['300']
                 : Colors.white['50'],
         },
     });
 
 /**
- * @param text  The text shown inside of the instruction
- * @param style  (Optional) Custom style to style the instruction.
- * @param isOutlineOnly  (Optional) If true, the buttom will not have a background colour, but an outline.
+ * @param text  The text shown inside of the button
+ * @param style  (Optional) Custom style to style the button.
+ * @param outlined  (Optional) If true, the button will not have a background color, but an outline.
  * @param disabled  (Optional) If true, the button will be disabled.
  * @param onPress  Action to take when button is tapped.
+ * @param theme (Optional) react-native-paper theme partial to style the component.
  */
-type ToggleButtonProps = {
+export type ToggleButtonProps = {
     text: string;
     style?: StyleProp<ViewStyle>;
-    isOutlineOnly?: boolean;
+    outlined?: boolean;
     disabled?: boolean;
-    onPress: Function;
+    onPress: () => void;
+    theme?: Theme;
 };
 
 /**
@@ -56,21 +57,21 @@ type ToggleButtonProps = {
  *
  * @category Component
  */
-export const ToggleButton = (props: ToggleButtonProps): JSX.Element => {
-    const isDisabled = props.disabled ?? false;
-    const isOutlineOnly = props.isOutlineOnly ?? false;
-    const styles = makeStyles(isOutlineOnly, isDisabled);
+export const ToggleButton: React.FC<ToggleButtonProps> = (props) => {
+    const { text, style, outlined = false, disabled = false, onPress, theme: customTheme } = props;
+    const theme = useTheme(customTheme);
+    const styles = makeStyles(props, theme);
 
     return (
         <Button
             uppercase={false}
-            mode={props.isOutlineOnly ? 'outlined' : 'contained'}
-            style={[styles.loginButton, props.style]}
-            disabled={isDisabled}
-            onPress={(): void => props.onPress()}
+            mode={outlined ? 'outlined' : 'contained'}
+            style={[styles.loginButton, style]}
+            disabled={disabled}
+            onPress={onPress}
         >
             <Label color="text" style={styles.label}>
-                {props.text}
+                {text}
             </Label>
         </Button>
     );