+++ login-workflow/src/screens/ResetPassword/ResetPasswordNav.tsx
@@ -13,14 +13,14 @@ import { useLanguageLocale } from '../../hooks/language-locale-hooks';
 import { useRoute } from '@react-navigation/native';
 
 // Screens
-import ResetPassword from '../../subScreens/ResetPassword';
-import ResetPasswordSent from '../../subScreens/ResetPasswordSent';
+import { ResetPassword } from '../../subScreens/ResetPassword';
+import { ResetPasswordSent } from '../../subScreens/ResetPasswordSent';
 
 // Components
-import CloseHeader from '../../components/CloseHeader';
+import { CloseHeader } from '../../components/CloseHeader';
 
 // Theme
-import { blue as BlueTheme } from '@pxblue/react-native-themes';
+import { Theme, useTheme } from 'react-native-paper';
 import { useAccountUIState, useAccountUIActions, AccountActions } from '../../contexts/AccountUIContext';
 
 // Types
@@ -32,24 +32,31 @@ import { ContactParams } from '../../types/ContactParams';
 const Stack = createStackNavigator();
 
 /**
+ * @param theme  (Optional) react-native-paper theme partial to style the component.
+ */
+type ResetPasswordNavProps = {
+    theme?: Theme;
+};
+
+/**
  * Renders a screen stack which handles the reset password flow.
  *
  * @category Component
  */
-function ResetPasswordNav(): JSX.Element {
+export const ResetPasswordNav: React.FC<ResetPasswordNavProps> = (props) => {
     const { t } = useLanguageLocale();
+    const theme = useTheme(props.theme);
     const accountUIState = useAccountUIState();
     const accountUIActions = useAccountUIActions();
     const route = useRoute();
 
     // Reset state on dismissal
-    // eslint-disable-next-line arrow-body-style
-    React.useEffect(() => {
-        return (): void => {
+    React.useEffect(
+        () => (): void => {
             accountUIActions.dispatch(AccountActions.resetPasswordReset());
-        };
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, []);
+        }, // eslint-disable-next-line react-hooks/exhaustive-deps
+        []
+    );
 
     const resetPassword = (emailInput: string): void => {
         accountUIActions.actions.forgotPassword(emailInput);
@@ -61,8 +68,8 @@ function ResetPasswordNav(): JSX.Element {
     return (
         <Stack.Navigator
             screenOptions={{
-                headerTintColor: BlueTheme.colors.onPrimary,
-                headerStyle: { backgroundColor: BlueTheme.colors.primary },
+                headerTintColor: theme.colors.surface,
+                headerStyle: { backgroundColor: theme.colors.primary },
             }}
         >
             {accountUIState.forgotPassword.transitSuccess !== true ? (
@@ -71,7 +78,7 @@ function ResetPasswordNav(): JSX.Element {
                     initialParams={{ onResetPasswordPress: resetPassword, contactPhone: contactPhone }}
                     component={ResetPassword}
                     options={({ navigation }): object => ({
-                        header: (): JSX.Element =>
+                        header: (): JSX.Element | null =>
                             CloseHeader({
                                 title: t('FORMS.RESET_PASSWORD'),
                                 backAction: () => {
@@ -87,7 +94,7 @@ function ResetPasswordNav(): JSX.Element {
                     initialParams={{ email: accountUIState.forgotPassword.email }}
                     options={({ navigation }): object => ({
                         gestureEnabled: false,
-                        header: (): JSX.Element =>
+                        header: (): JSX.Element | null =>
                             CloseHeader({
                                 title: t('FORMS.RESET_PASSWORD'),
                                 backAction: () => {
@@ -99,6 +106,4 @@ function ResetPasswordNav(): JSX.Element {
             )}
         </Stack.Navigator>
     );
-}
-
-export default ResetPasswordNav;
+};