+++ login-workflow/src/subScreens/ResetPasswordSent.tsx
@@ -8,11 +8,9 @@ import React from 'react';
 // Components
 import { View, StyleSheet, SafeAreaView } from 'react-native';
 import { Instruction } from '../components/Instruction';
-import { Icon } from 'react-native-elements';
+import { Theme, useTheme } from 'react-native-paper';
+import MatIcon from 'react-native-vector-icons/MaterialIcons';
 import { ToggleButton } from '../components/ToggleButton';
-
-// Styles
-import * as Colors from '@pxblue/colors';
 
 // Hooks
 import { useLanguageLocale } from '../hooks/language-locale-hooks';
@@ -24,12 +22,11 @@ import { ResetPasswordParams } from '../types/ResetPasswordParams';
 /**
  * @ignore
  */
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-const makeContainerStyles = () =>
+const makeContainerStyles = (theme: Theme): Record<string, any> =>
     StyleSheet.create({
         safeContainer: {
             height: '100%',
-            backgroundColor: Colors.white['50'],
+            backgroundColor: theme.colors.surface,
             flex: 1,
             justifyContent: 'space-between',
         },
@@ -42,14 +39,14 @@ const makeContainerStyles = () =>
         iconContainer: {
             marginTop: 60,
             marginBottom: 20,
+            alignSelf: 'center',
         },
     });
 
 /**
  * @ignore
  */
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-const makeStyles = () =>
+const makeStyles = (): Record<string, any> =>
     StyleSheet.create({
         wideButton: {
             height: 60,
@@ -58,16 +55,24 @@ const makeStyles = () =>
     });
 
 /**
+ * @param theme (Optional) react-native-paper theme partial to style the component.
+ */
+type ResetPasswordSentProps = {
+    theme?: Theme;
+};
+
+/**
  * Renders the screen with the reset password sent message.
  *
  * @category Component
  */
-export default function ResetPasswordSent(): JSX.Element {
+export const ResetPasswordSent: React.FC<ResetPasswordSentProps> = (props) => {
+    const theme = useTheme(props.theme);
     const { t } = useLanguageLocale();
     const navigation = useNavigation();
     const route = useRoute();
 
-    const containerStyles = makeContainerStyles();
+    const containerStyles = makeContainerStyles(theme);
     const styles = makeStyles();
 
     const routeParams = route.params as ResetPasswordParams;
@@ -76,11 +81,11 @@ export default function ResetPasswordSent(): JSX.Element {
     return (
         <SafeAreaView style={containerStyles.safeContainer}>
             <View>
-                <Icon
+                <MatIcon
                     name={'mail-outline'}
-                    containerStyle={containerStyles.iconContainer}
+                    style={containerStyles.iconContainer}
                     size={100}
-                    color={Colors.gray['500']}
+                    color={theme.colors.placeholder}
                 />
 
                 <Instruction
@@ -99,4 +104,4 @@ export default function ResetPasswordSent(): JSX.Element {
             </View>
         </SafeAreaView>
     );
-}
+};